Move upgrade version into automation
