+++ ci/ray_ci/automation/BUILD.bazel
@@ -1,4 +1,5 @@
 load("@rules_python//python:defs.bzl", "py_binary", "py_library", "py_test")
+load("@py_deps_buildkite//:requirements.bzl", ci_require = "requirement")
 
 py_binary(
     name = "state_machine_bot",
@@ -24,3 +25,28 @@ py_binary(
     deps = ["//ci/ray_ci:ray_ci_lib"],
 )
 
+py_library(
+    name = "upgrade_version_lib",
+    srcs = ["upgrade_version_lib.py"],
+    visibility = ["//ci/ray_ci/automation:__subpackages__"],
+    deps = [],
+)
+
+py_test(
+    name = "test_upgrade_version_lib",
+    size = "small",
+    srcs = ["test_upgrade_version_lib.py"],
+    tags = [
+        "ci_unit",
+        "team:ci",
+    ],
+    deps = [
+        ci_require("pytest"),
+        ":upgrade_version_lib",
+    ],
+)
+
+py_binary(
+    name = "upgrade_version",
+    srcs = ["upgrade_version.py"],
+)
